// src/pages/ServicesPage/types.ts
export enum EServices {
  WEB_AND_MOB_DEV = 'Web and Mobile development',
  DEVOPS_SOLUTIONS = 'DevOps solutions',
  ARCHITECTURE_DESIGN = 'Architecture Design',
  POC_PROT_MVP = 'POC, Prototype, MVP',
  TESTIMG = 'QA & Testing',
  DESIGN = 'UX/UI Design',
}

export enum EServicesIDs {
  WEB_AND_MOB_DEV = 'webMobDev',
  DEVOPS_SOLUTIONS = 'devops',
  ARCHITECTURE_DESIGN = 'archDesign',
  POC_PROT_MVP = 'pocProtMVP',
  TESTIMG = 'testing',
  DESIGN = 'design',
}